Separate ion token refresh from CesiumIonTilesetLoader.
